requirements.txt
anthropic>=0.40.0
fredapi>=0.5.1
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
dash>=2.0.0
plotly>=5.0.0
pandas
numpy
gunicorn